Split service files into folders and replaced framer motion with css animations, resulting in 74% fewer modules (from 945 to 245), and a 112KB reduction in bundle size
