+++ package.json
@@ -16,7 +16,6 @@
     "firebase-admin": "^13.5.0",
     "firebase-functions": "^6.3.2",
     "firebase-tools": "^14.15.2",
-    "framer-motion": "^12.10.5",
     "papaparse": "^5.5.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
