+++ src/components/flashcards/MyFlashcards.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { useAuth } from '../../contexts/AuthContext';
 import { Link } from 'react-router-dom';
-import flashcardService from '../../services/flashcardService';
+import flashcardService from '../../services/flashcards/flashcardService';
 
 export default function MyFlashcards() {
   const { currentUser } = useAuth();
