+++ src/services/quiz/resultService.js
@@ -2,7 +2,7 @@
  * Result Service - handles basic quiz result retrieval
  */
 import { collection, doc, getDoc, getDocs, query, where, orderBy, limit, startAfter } from 'firebase/firestore';
-import { db, handleFirebaseError, withRetry, timestamp } from './firebaseService';
+import { db, handleFirebaseError, withRetry, timestamp } from '../firebase/firebaseService';
 
 class ResultService {
     constructor() {
