+++ src/styles/utilities.css
@@ -66,4 +66,92 @@
   .btn-hover:disabled:hover {
     transform: none;
   }
+
+  /* Keyframe animations */
+  @keyframes fadeInUp {
+    from {
+      opacity: 0;
+      transform: translateY(30px);
+    }
+    to {
+      opacity: 1;
+      transform: translateY(0);
+    }
+  }
+
+  @keyframes fadeInUpLarge {
+    from {
+      opacity: 0;
+      transform: translateY(40px);
+    }
+    to {
+      opacity: 1;
+      transform: translateY(0);
+    }
+  }
+
+  @keyframes fadeInRotate {
+    from {
+      opacity: 0;
+      transform: rotateX(-10deg);
+    }
+    to {
+      opacity: 1;
+      transform: rotateX(0deg);
+    }
+  }
+
+  /* Animation classes */
+  .animate-fade-in-up {
+    animation: fadeInUp 0.6s ease-out forwards;
+  }
+
+  .animate-fade-in-up-large {
+    animation: fadeInUpLarge 1s ease-out forwards;
+  }
+
+  .animate-fade-in-rotate {
+    animation: fadeInRotate 1s ease-out forwards;
+  }
+
+  /* Staggered animations with delays */
+  .animate-fade-in-up-delay-100 {
+    animation: fadeInUp 0.6s ease-out 0.1s forwards;
+    opacity: 0;
+  }
+
+  .animate-fade-in-up-delay-200 {
+    animation: fadeInUp 0.6s ease-out 0.2s forwards;
+    opacity: 0;
+  }
+
+  .animate-fade-in-up-delay-300 {
+    animation: fadeInUp 0.6s ease-out 0.3s forwards;
+    opacity: 0;
+  }
+
+  /* Interactive hover effects */
+  .animate-hover-scale {
+    transition: transform 0.2s ease;
+  }
+
+  .animate-hover-scale:hover {
+    transform: scale(1.03);
+  }
+
+  .animate-tap-scale:active {
+    transform: scale(0.97);
+  }
+
+  /* Scroll-triggered animations (for elements that should animate when in view) */
+  .animate-on-scroll {
+    opacity: 0;
+    transform: translateY(20px);
+    transition: opacity 0.6s ease-out, transform 0.6s ease-out;
+  }
+
+  .animate-on-scroll.animate-in {
+    opacity: 1;
+    transform: translateY(0);
+  }
 }
